requirements.txt
Django>=4.2
djangorestframework
mysqlclient
pillow
bleach
celery
redis
django-cors-headers
whitenoise
django-simple-captcha
django-cors-headers
djangorestframework-simplejwt
django-filter>=21.1
channels
daphne
channels_redis
